+++ rosomaxa/src/termination/target_proximity.rs
@@ -45,7 +45,7 @@ where
 
     fn is_termination(&self, heuristic_ctx: &mut Self::Context) -> bool {
         // NOTE ignore pareto front, use the first solution only for comparison
-        heuristic_ctx.population().ranked().next().map_or(false, |(solution, _)| {
+        heuristic_ctx.ranked().next().map_or(false, |(solution, _)| {
             let distance = relative_distance(self.target_fitness.iter().cloned(), solution.fitness());
             distance < self.distance_threshold
         })