+++ vrp-core/src/solver/search/decompose_search.rs
@@ -79,7 +79,7 @@ impl DecomposeSearch {
         // do actual refinement independently for each decomposed context
         let decomposed = parallel_into_collect(decomposed, |(mut refinement_ctx, route_indices)| {
             let _ = (0..self.repeat_count).try_for_each(|_| {
-                let insertion_ctx = refinement_ctx.population().select().next().expect(GREEDY_ERROR);
+                let insertion_ctx = refinement_ctx.selected().next().expect(GREEDY_ERROR);
                 let insertion_ctx = self.inner_search.search(&refinement_ctx, insertion_ctx);
                 let is_quota_reached =
                     refinement_ctx.environment.quota.as_ref().map_or(false, |quota| quota.is_reached());
@@ -271,7 +271,7 @@ fn merge_best(
     accumulated: InsertionContext,
 ) -> InsertionContext {
     let (decomposed_ctx, route_indices) = decomposed;
-    let (decomposed_insertion_ctx, _) = decomposed_ctx.population().ranked().next().expect(GREEDY_ERROR);
+    let (decomposed_insertion_ctx, _) = decomposed_ctx.ranked().next().expect(GREEDY_ERROR);
     let environment = original_insertion_ctx.environment.clone();
 
     let (partial_insertion_ctx, _) = create_partial_insertion_ctx(original_insertion_ctx, environment, route_indices);